+++ server/src/main/java/com/codestates/server/domain/member/dto/MemberPostDto.java
@@ -1,4 +1,4 @@
-package com.codestates.server.member.dto;
+package com.codestates.server.domain.member.dto;
 
 import lombok.Getter;
 import lombok.Setter;
