+++ server/src/main/java/com/codestates/server/domain/member/mapper/MemberMapper.java
@@ -1,10 +1,10 @@
-package com.codestates.server.member.mapper;
+package com.codestates.server.domain.member.mapper;
 
-import com.codestates.server.member.dto.MemberLoginResponseDto;
-import com.codestates.server.member.dto.MemberPatchDto;
-import com.codestates.server.member.dto.MemberPostDto;
-import com.codestates.server.member.dto.MemberResponseDto;
-import com.codestates.server.member.entity.Member;
+import com.codestates.server.domain.member.dto.MemberLoginResponseDto;
+import com.codestates.server.domain.member.dto.MemberPatchDto;
+import com.codestates.server.domain.member.dto.MemberPostDto;
+import com.codestates.server.domain.member.dto.MemberResponseDto;
+import com.codestates.server.domain.member.entity.Member;
 import org.mapstruct.Mapper;
 
 import java.util.List;
